REFINE: Introduce AsyncToken class for simplified token handling in AsyncStorage
